fix(assets): organize assets
