--- src/assets/icons/LabsIcon.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { SvgIcon } from '@material-ui/core';
-
-function LabsIcon() {
-  return (
-    <SvgIcon>
-      <path d="M33.99 28C31.195 28 29 25.799 29 23c0-2.797 2.194-5 4.99-5 2.798 0 5.01 2.203 5.01 5 0 2.799-2.212 5-5.01 5zm0-8.656c-1.987 0-3.523 1.58-3.523 3.657s1.536 3.655 3.524 3.655 3.537-1.598 3.537-3.655c0-2.058-1.536-3.657-3.537-3.657zM40.005 21.11h1.348v.678c.59-.52 1.35-.8 2.133-.788C45.458 21 47 22.594 47 24.62c0 2.024-1.545 3.631-3.541 3.631-.753.002-1.486-.24-2.092-.69V31H40l.005-9.89zm1.367 1.934v3.177c.438.51 1.162.828 1.942.828 1.34 0 2.352-1.036 2.352-2.417 0-1.38-1.026-2.43-2.352-2.43-.793 0-1.49.304-1.942.842zM51.73 27.999c-2.095 0-3.73-1.536-3.73-3.512C48 22.53 49.565 21 51.562 21 53.518 21 55 22.543 55 24.54v.383h-5.615c.179 1.129 1.205 1.958 2.403 1.942.655.014 1.293-.196 1.8-.59l.895.84c-.78.594-1.755.907-2.753.884zm-2.333-4.072h4.22c-.21-1.037-1.048-1.81-2.083-1.81-1.074 0-1.927.733-2.137 1.81zM56 21.136h1.327v.687c.538-.545 1.272-.842 2.032-.822C60.912 21 62 22.093 62 23.685V28h-1.327v-4.09c0-1.078-.65-1.74-1.673-1.74-.672-.022-1.307.316-1.672.89v4.936h-1.327L56 21.136zM67 19L68 19 68 28 67 28zM70 21.136h1.327v.687c.539-.544 1.272-.841 2.031-.822C74.911 21 76 22.093 76 23.685V28h-1.327v-4.09c0-1.078-.651-1.74-1.674-1.74-.672-.022-1.307.316-1.672.89v4.936H70v-6.86zM77 21.136h1.328v.687c.538-.544 1.272-.841 2.03-.822C81.913 21 83 22.093 83 23.685V28h-1.328v-4.09c0-1.078-.65-1.74-1.672-1.74-.673-.022-1.308.316-1.672.89v4.936H77v-6.86zM87.506 21c1.416.003 2.691.858 3.23 2.167.54 1.31.237 2.814-.767 3.813-1.003 1-2.51 1.295-3.816.75-1.307-.545-2.156-1.823-2.153-3.239-.009-.931.358-1.826 1.018-2.484.66-.657 1.557-1.02 2.488-1.007zm2.198 3.491c0-1.31-.98-2.33-2.198-2.33-1.217 0-2.21 1.019-2.21 2.33 0 1.312.966 2.342 2.21 2.342 1.218.003 2.198-1.03 2.198-2.342zM92 21L93.439 21 95.52 26.24 97.601 21 99 21 96.121 28 94.879 28zM99 25.94c0-1.273 1.057-2.053 2.789-2.053.647 0 1.288.126 1.885.372v-.716c0-.954-.583-1.43-1.679-1.43-.636 0-1.285.172-2.137.57l-.501-.994c1.028-.478 1.921-.689 2.828-.689 1.786 0 2.815.861 2.815 2.411v4.458h-1.326v-.572c-.605.475-1.364.724-2.14.702-1.503-.005-2.534-.84-2.534-2.059zm2.843 1.074c.668.02 1.318-.208 1.822-.637v-1.166c-.545-.284-1.156-.426-1.774-.41-.987 0-1.596.436-1.596 1.113.005.65.627 1.1 1.548 1.1zM107.512 22.19H106v-1.126h1.512v-1.742L108.9 19v2.062H111v1.126h-2.1v3.695c0 .697.293.952 1.05.952.348.009.695-.046 1.022-.16v1.124c-.431.129-.88.197-1.33.201-1.388 0-2.13-.629-2.13-1.804V22.19zM111 18.867c-.002-.228.1-.447.282-.61.182-.162.43-.255.69-.256.362-.013.704.15.889.423.185.273.185.614 0 .887s-.527.436-.889.423c-.26-.002-.509-.094-.691-.257-.182-.163-.283-.382-.281-.61zM112.77 28h-1.595v-7.119h1.594V28zM117.506 21c1.416.003 2.691.858 3.23 2.167.54 1.31.237 2.814-.767 3.813-1.003 1-2.51 1.295-3.816.75-1.307-.545-2.156-1.823-2.153-3.239-.009-.931.358-1.826 1.018-2.484.66-.657 1.557-1.02 2.488-1.007zm2.198 3.491c0-1.31-.98-2.33-2.198-2.33-1.217 0-2.204 1.022-2.204 2.33s.967 2.342 2.21 2.342c1.212.003 2.192-1.03 2.192-2.342zM123 21.136h1.327v.687c.538-.545 1.273-.842 2.032-.822 1.553 0 2.641 1.092 2.641 2.684V28h-1.327v-4.09c0-1.078-.65-1.74-1.672-1.74-.673-.022-1.307.316-1.672.89v4.936h-1.326l-.003-6.86zM133 19L134.447 19 134.447 26.805 140 26.805 140 28 133 28zM141 25.94c0-1.273 1.056-2.053 2.79-2.053.647 0 1.286.127 1.882.372v-.716c0-.954-.582-1.43-1.679-1.43-.636 0-1.286.172-2.14.57l-.501-.994c1.03-.478 1.923-.689 2.83-.689 1.788 0 2.818.861 2.818 2.411v4.458h-1.328v-.572c-.605.475-1.364.724-2.14.702-1.502-.005-2.532-.84-2.532-2.059zm2.844 1.074c.668.02 1.319-.208 1.823-.637v-1.166c-.545-.284-1.156-.426-1.774-.41-.988 0-1.597.436-1.597 1.113.005.65.627 1.1 1.548 1.1zM151.457 28c-.758.003-1.497-.242-2.103-.698v.589H148V18.3l1.367-.301v3.575c.592-.502 1.344-.775 2.12-.767 1.971 0 3.513 1.58 3.513 3.59 0 2.008-1.547 3.602-3.543 3.602zm-2.086-5.165v3.151c.438.507 1.162.822 1.942.822 1.34 0 2.351-1.027 2.351-2.397 0-1.37-1.025-2.41-2.351-2.41-.8 0-1.496.3-1.942.834zM156.745 26.121c.73.56 1.5.853 2.284.853.991 0 1.68-.414 1.68-1.013 0-.478-.36-.758-1.13-.865l-1.266-.172c-1.39-.2-2.092-.826-2.092-1.876 0-1.225 1.101-2.048 2.74-2.048 1.01-.006 1.998.297 2.82.865l-.7.878c-.772-.491-1.474-.705-2.217-.705-.84 0-1.404.372-1.404.919 0 .492.33.732 1.156.85l1.267.175c1.39.2 2.117.838 2.117 1.889 0 1.211-1.265 2.129-2.931 2.129-1.211 0-2.299-.346-3.069-.998l.745-.88zM16.68 10.951c1.643 0 4.02-.34 4.02-2.303.005-.152-.009-.303-.04-.452l-.98-4.27c-.225-.939-.423-1.365-2.065-2.19C16.337 1.081 13.562 0 12.74 0c-.765 0-.992.995-1.898.995-.877 0-1.529-.739-2.35-.739-.788 0-1.301.54-1.698 1.65 0 0-1.104 3.129-1.246 3.583-.024.083-.034.17-.032.256 0 1.216 4.772 5.206 11.164 5.206m4.275-1.506c.227 1.08.227 1.194.227 1.336 0 1.847-2.067 2.873-4.785 2.873C10.244 13.652 4.87 10.04 4.87 7.653c0-.333.067-.662.198-.967C2.86 6.798 0 7.196 0 9.726 0 13.882 9.796 19 17.553 19 23.5 19 25 16.298 25 14.164c0-1.678-1.445-3.583-4.05-4.72M74.781 11.704c0 1.517.925 2.255 2.613 2.255.52-.013 1.036-.085 1.539-.214v-1.759c-.322.1-.657.15-.994.148-.695 0-.953-.214-.953-.859V8.577H79V6.765h-2.014V4.467l-2.2.47v1.828h-1.46v1.812h1.456l-.001 3.127zm-6.859.04c0-.469.478-.697 1.198-.697.441.002.88.056 1.307.16v.913c-.42.226-.892.342-1.371.336-.707 0-1.13-.268-1.13-.711m.673 2.241c.776 0 1.403-.16 1.988-.55v.43h2.176V9.315c0-1.731-1.182-2.68-3.156-2.68-1.103 0-2.192.255-3.365.778l.79 1.598c.843-.35 1.553-.564 2.178-.564.906 0 1.374.348 1.374 1.06v.349c-.533-.136-1.082-.204-1.633-.202-1.852 0-2.968.766-2.968 2.135 0 1.249 1.007 2.2 2.613 2.2m-11.963-.12h2.341v-3.681h3.92v3.678h2.341V4.47H62.9v3.61h-3.92V4.47h-2.34l-.006 9.4zm-8.916-3.56c0-1.021.817-1.8 1.893-1.8.56-.012 1.104.185 1.524.552v2.477c-.412.38-.96.584-1.524.564-1.062 0-1.893-.778-1.893-1.798m3.444 3.557h2.178V4l-2.2.47V7.14c-.56-.313-1.194-.475-1.837-.47-2.096 0-3.743 1.597-3.743 3.638-.013.967.368 1.899 1.058 2.586.69.687 1.631 1.072 2.612 1.068.694.003 1.371-.213 1.932-.616v.514zm-9.99-5.45c.694 0 1.28.443 1.51 1.127h-3.003c.218-.71.762-1.127 1.497-1.127m-3.719 1.91c0 2.068 1.715 3.679 3.919 3.679 1.211 0 2.096-.323 3.009-1.075l-1.458-1.276c-.34.35-.843.537-1.441.537-.77.022-1.473-.426-1.77-1.127h5.13v-.536c0-2.255-1.537-3.878-3.633-3.878-.996-.015-1.956.367-2.662 1.059s-1.1 1.636-1.091 2.618m-3.8-3.88c.776 0 1.211.483 1.211 1.062 0 .578-.436 1.06-1.211 1.06h-2.314V6.443h2.314zM29 13.867h2.341v-3.423h1.782l1.798 3.423h2.613l-2.097-3.759c1.09-.437 1.801-1.484 1.796-2.644 0-1.691-1.347-2.993-3.364-2.993h-4.865L29 13.866z" />
-    </SvgIcon>
-  );
-}
-
-export default LabsIcon;
