--- src/assets/icons/NewTabIcon.js
@@ -1,19 +1,0 @@
-import React from "react";
-
-function NewTabIcon(props) {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg"
-      width={props.width || "11"}
-      height={props.height || "11"}
-      viewBox="0 0 11 11"
-    >
-      <g fill={props.fill || "#1975FF"}>
-        <path d="M9.77 5.688v2.561c0 .17.005.342-.018.511l.023-.163c-.024.164-.067.323-.13.477l.06-.147c-.063.15-.144.29-.243.417l.096-.124c-.098.126-.212.24-.339.34l.124-.097c-.129.099-.268.18-.417.244l.147-.062c-.153.064-.312.107-.477.13l.163-.022c-.257.033-.525.017-.783.017H3.163c-.224 0-.447.002-.67 0-.086 0-.17-.006-.255-.017l.163.022c-.165-.023-.323-.066-.477-.13l.146.062c-.148-.064-.288-.145-.416-.244l.124.096c-.127-.098-.241-.213-.34-.34l.096.125c-.098-.129-.18-.268-.243-.417l.062.147c-.064-.153-.107-.313-.13-.477l.021.163c-.033-.257-.017-.525-.017-.783v-1.47-1.804-1.538-.67c0-.086.006-.171.017-.256l-.022.164c.024-.165.067-.324.13-.477l-.06.146c.063-.149.144-.289.242-.417l-.095.125c.098-.127.212-.241.339-.34l-.124.096c.129-.098.268-.18.416-.243l-.146.061c.153-.064.312-.107.477-.13l-.163.022c.148-.02.298-.017.448-.017H5.31c.158 0 .324-.069.435-.18.107-.106.187-.28.18-.435-.007-.158-.06-.323-.18-.435C5.625.07 5.478 0 5.31 0H2.504c-.403 0-.788.098-1.148.275C.953.473.613.803.374 1.18c-.246.388-.372.855-.373 1.314-.002.172 0 .344 0 .517V8.506c0 .434.115.869.331 1.245.44.767 1.281 1.246 2.164 1.248.19.002.38 0 .571 0H8.377c.054 0 .11.002.164 0 .458-.008.929-.14 1.312-.395.373-.247.698-.589.889-.997.165-.355.258-.723.258-1.116V7.397 5.946v-.257c0-.157-.069-.323-.18-.435-.106-.107-.28-.187-.435-.18-.158.008-.323.06-.435.18-.11.12-.18.267-.18.434zM7.72 1.294h2.677c.154 0 .317-.072.427-.189.104-.112.183-.296.176-.458-.008-.167-.058-.34-.176-.458C10.705.072 10.56 0 10.397 0H7.72c-.153 0-.316.072-.426.189-.105.113-.183.296-.176.458.007.167.058.34.176.458.118.115.264.19.427.19z"/>
-        <path d="M11 3.28v-.906V.932.603c0-.154-.072-.317-.189-.427-.113-.104-.296-.183-.458-.176-.167.008-.34.058-.458.176-.117.119-.19.263-.19.427V3.28c0 .153.073.316.19.426.112.105.296.183.458.176.167-.007.34-.058.458-.176.115-.118.189-.264.189-.427z"/>
-        <path d="M6.178 5.652l.465-.465 1.11-1.11 1.34-1.34 1.164-1.163c.187-.188.378-.374.564-.564l.008-.008C10.936.895 11 .738 11 .587c0-.145-.063-.316-.171-.415-.112-.102-.258-.179-.415-.172-.156.007-.303.06-.415.172l-.466.465-1.109 1.109-1.341 1.341L5.92 4.25c-.188.187-.379.374-.564.564l-.008.008c-.107.107-.172.264-.172.415 0 .144.064.315.172.415.111.102.258.178.415.171.156-.006.302-.059.415-.171z"/>
-      </g>
-    </svg>
-  );
-}
-
-export default NewTabIcon;
