--- src/assets/icons/RockIcon.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { SvgIcon } from '@material-ui/core';
-
-function RockIcon(props) {
-  return (
-    <SvgIcon viewBox="-3 -3 22 22">
-      <path d="M9.58 1.767c-.682 0-1.235.553-1.235 1.235v1.182c-.509-.283-1.147-.172-1.529.268-.385-.42-1.008-.52-1.505-.242V2.274c0-.682-.553-1.235-1.235-1.235-.681 0-1.234.553-1.234 1.235V5.61c-.716.098-1.242.51-1.529 1.206-.248.602-.248 1.242-.248 1.516v.217c0 2.802 2.006 4.759 4.877 4.759 1.775 0 3.093-.602 3.915-1.789.634-.92.958-2.176.958-3.729V3.006c.001-.328-.129-.644-.36-.876-.232-.233-.547-.363-.876-.363zM5.937 12.613c-1.234 0-2.294-.4-3.051-1.158-.73-.728-1.13-1.76-1.13-2.906V8.33c0-1.201.354-1.865 1.08-2.016v1.15c0 .192.155.347.347.347.191 0 .347-.155.347-.347v-5.19c0-.298.242-.54.54-.54.298 0 .54.242.54.54v3.234c0 .192.156.348.347.348h.059c.192 0 .347-.156.347-.348v-.217c0-.298.242-.54.54-.54.299 0 .54.242.54.54v.217c0 .192.156.348.348.348h.059c.191 0 .347-.156.347-.348v-.243c0-.298.241-.54.54-.54.298 0 .54.242.54.54v.243c0 .192.155.348.347.348h.059c.191 0 .347-.156.347-.348V3.006c0-.298.242-.54.54-.54.298 0 .54.242.54.54V7.79c.01 3.2-1.395 4.823-4.173 4.823z" />
-    </SvgIcon>
-  );
-}
-
-export default RockIcon;
