+++ src/assets/icons/index.js
@@ -3,15 +3,10 @@ export { default as AmaIcon } from './AmaIcon';
 export { default as FilledHeartIcon } from './FilledHeartIcon';
 export { default as UserIcon } from './UserIcon';
 export { default as CameraIcon } from './CameraIcon';
-export { default as RockIcon } from './RockIcon';
 export { default as CCHeartIcon } from './CCHeartIcon';
-export { default as BetaIcon } from './BetaIcon';
-export { default as NewTabIcon } from './NewTabIcon';
 export { default as WebLinkIcon } from './WebLinkIcon';
 export { default as PodcastIcon } from './PodcastIcon';
 export { default as DownloadIcon } from './DownloadIcon';
-export { default as LabsIcon } from './LabsIcon';
-export { default as ArrowDownIcon } from './ArrowDownIcon';
 export { default as ArrowRightIcon } from './ArrowRightIcon';
 export { default as OPLLogo } from './OPLLogo';
 export { default as EditIcon } from './EditIcon';
